Move EOS to utils (to mimic ERF source)
